fix unlink temp file #4

+++ lib/relaton/index/file_io.rb
@@ -130,8 +130,8 @@ module Relaton
       def fetch_and_save
         resp = URI(url).open
         zip = Zip::InputStream.new resp
-        resp.close
-        resp.unlink
+        resp.close if resp.respond_to? :close
+        resp.unlink if resp.is_a? Tempfile
         entry = zip.get_next_entry
         index = YAML.safe_load(entry.get_input_stream.read, permitted_classes: [Symbol])
         save index
